server/scripts/Python/requirements.txt
Flask
flask_pymongo
flask_cors
google-cloud-vision
google-api-core
openai
pymongo
